refactor: restructure header
