+++ src/pages/home-page/home-page.tsx
@@ -1,6 +1,6 @@
 import { Box } from '@mui/material';
 
-import { Header } from '@/shared/header';
+import { Header } from '@/layouts/header';
 
 import { useStyles } from './styles';
